Added event target for select example
